Add tracing script.
This commit adds a more advanced version of the Bash-based tracing
script that has been used thus far. It requires several command line
parameters for figuring out where stuff lives, but that's the only major
wrinkle.

It generates an LTTng tracing session around a user-specified command
(assumed to be an OpenMP-based program).